pkg/seq/dom/int/html: renames the package from `nodes` to better distinguish from potentially upcoming `svg` package
