+++ pkg/sequence/dom/internal/html/htmler.go
@@ -1,4 +1,4 @@
-package nodes
+package html
 
 import (
 	"fmt"
